+++ src/components/question/index.tsx
@@ -1,18 +1,23 @@
-import React, { useState } from 'react'
+import React, { useContext, useState } from 'react'
 import { Option } from '../option'
 import { chromeContext } from '../../global-styles'
+import { FormContext } from '../../App'
 
 interface QuesitonProps {
   title: string
   options: Option[]
 }
 
-interface Option {
+export interface Option {
   name: string
   weight: number
 }
 
 export const Question = ({ title, options }: QuesitonProps) => {
+  const {
+    form,
+    setForm
+  } = useContext(FormContext);
   const [currentlySelectedOption, setCurrentlySelectedOption] = useState(99999)
 
   return (
@@ -21,11 +26,17 @@ export const Question = ({ title, options }: QuesitonProps) => {
         {title}
       </header>
       <section style={styles.optionsContainer}>
-        {options.map((option, index) => {
+        {options.map((option) => {
           return (
             <Option
-              isSelected={index === currentlySelectedOption}
-              onClick={() => setCurrentlySelectedOption(index)}
+              isSelected={option.weight === currentlySelectedOption}
+              onClick={() => {
+                setCurrentlySelectedOption(option.weight);
+                setForm({
+                  ...form,
+                  [title]: option.weight
+                });
+              }}
               {...option}
             />
           )
